crawler/requirements.txt
apscheduler
uvicorn
fastapi
google-cloud-bigquery
pandas
requests
python-dotenv
pyarrow